Don’t override ⌘B if there is a selection (in text)
The motivation for this is that ⌘B in markup languages (which should have a text scope) are bound to “Format → Bold”, which is an action that primarily is used with a selection. So by scoping the Ninja build command to only the case where there is no selection (in the text scope), we can have both Bold and Build coexist in markup languages.

Since users may not realize that it is having a selection that stops ⌘B from working, I have limited this “feature” to the text scope, so in source files (where ⌘B for Build sees its primary use) we can still use ⌘B to build with a selection.